+++ lib/resdayn/parser/record/land_texture.ex
@@ -4,7 +4,7 @@ defmodule Resdayn.Parser.Record.LandTexture do
   process_basic_string "NAME", :id
   process_basic_string "DATA", :texture
 
-  def process({"INTV", <<value::long()>>}, data) do
+  def process({"INTV", <<value::uint32()>>}, data) do
     record_value(data, :index, value)
   end
 end
