+++ lib/resdayn/parser/record/script.ex
@@ -4,8 +4,8 @@ defmodule Resdayn.Parser.Record.Script do
   process_basic_string "SCTX", :text
 
   def process({"SCHD" = v, value}, data) do
-    <<name::char(32), num_shorts::long(), num_longs::long(), num_floats::long(),
-      data_size::long(), local_variable_size::long()>> = value
+    <<name::char(32), num_shorts::uint32(), num_longs::uint32(), num_floats::uint32(),
+      data_size::uint32(), local_variable_size::uint32()>> = value
 
     record_unnested_value(data, %{
       name: printable!(__MODULE__, v, name),
